requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.7
python-docx==1.0.1
pdfminer.six==20231228
aiofiles==23.2.1
Jinja2==3.1.2
aiofiles==23.2.1
pydantic
PyPDF2